Tue, Aug 2, 2022, 9:19 AM +00:00

+++ Week 2/readability/readability.c
@@ -13,10 +13,10 @@ int main(void)
     int words = count_words(text);
     int sentences = count_sentences(text);
 
-    int L = (letters / words) * 100;
-    int S = (sentences / words) * 100;
+    float L = (letters / words) * 100;
+    float S = (sentences / words) * 100;
 
-    int index = 0.0588 * L - 0.296 * S - 15.8;
+    float index = 0.0588 * L - 0.296 * S - 15.8;
 
     if (index < 1)
     {
@@ -25,7 +25,7 @@ int main(void)
     {
         printf("Grade 16+");
     } else {
-        printf("Grade %i\n", index);
+        printf("Grade %i\n", (int) index);
     }
 }
